refactor(circle): extraire CircleOrganizersList pour dédupliquer le bloc "Organisé par"
Le bloc listant les organisateurs (avatars + noms cliquables) était
dupliqué à 95% entre les deux pages Circle (publique + dashboard). Seule
différence : la publique vérifie isConnected avant d'autoriser le lien
vers le profil, le dashboard suppose la session.

Extraction en composant avec prop linkable : vient du contexte d'usage
(isConnected sur publique, toujours true sur dashboard).

Bénéfice secondaire : grâce au fix S4 (backfill publicId assuré), la
vérification `host.user.publicId ? Link : div` disparaît aussi — une
source de fallback obsolète en moins.
